refactor code, naming, & readme info
