+++ tests/conftest.py
@@ -3,46 +3,17 @@ from typing import Generator
 
 import pytest
 
+from socialetl.schema_manager import setup_db_schema, teardown_db_schema
 from socialetl.utils.db import DatabaseConnection
 
 
 @pytest.fixture(scope='session', autouse=True)
-def mock_social_posts_table() -> Generator:
-    db = DatabaseConnection(db_file="data/test.db")
-    # Setup schema
-    with db.managed_cursor() as cur:
-        # setup schema
-        cur.execute(
-            """
-            CREATE TABLE IF NOT EXISTS social_posts (
-                id TEXT PRIMARY KEY,
-                source TEXT,
-                social_data TEXT,
-                dt_created datetime default current_timestamp
-            )
-            """
-        )
-        cur.execute(
-            """
-            CREATE TABLE IF NOT EXISTS log_metadata (
-                dt_created datetime default current_timestamp,
-                function_name TEXT,
-                input_params TEXT
-            )
-            """
-        )
+def mock_social_posts_table(session_mocker) -> Generator:
+    session_mocker.patch(
+        "socialetl.schema_manager.db_factory",
+        return_value=DatabaseConnection(db_file="data/test.db"),
+    )
+    setup_db_schema()
     yield
-    # teardown schema
-    with db.managed_cursor() as cur:
-        # setup schema
-        cur.execute(
-            """
-            DROP TABLE IF EXISTS social_posts
-            """
-        )
-        cur.execute(
-            """
-            DROP TABLE IF EXISTS log_metadata
-            """
-        )
+    teardown_db_schema()
     os.remove("data/test.db")
