+++ tests/test_twitter.py
@@ -3,8 +3,8 @@ from typing import List
 
 import pytest
 
-from socialetl.socialetl import (SocialMediaData, TransformationType,
-                                 TwitterTweetData, etl_factory)
+from socialetl.social_etl import SocialMediaData, TwitterTweetData, etl_factory
+from socialetl.transform import transformation_factory
 from socialetl.utils.db import DatabaseConnection
 
 
@@ -44,7 +44,7 @@ class TestTwitterETL:
             objects that replicate what we get from the extract method.
         """
         mocker.patch(
-            "socialetl.socialetl.db_factory",
+            "socialetl.social_etl.db_factory",
             return_value=DatabaseConnection(db_file="data/test.db"),
         )
         # Create a TwitterETL object
@@ -52,8 +52,7 @@ class TestTwitterETL:
         # Call the transform method on the TwitterETL object
         # and pass in the mock
         transformed_data = twitter_etl.transform(
-            mock_twitter_data,
-            TransformationType('no_tx').transformation_factory(),
+            mock_twitter_data, transformation_factory('no_tx')
         )
         # Assert that the transformed data is a
         # list of TwitterTweetData objects
@@ -91,20 +90,16 @@ class TestTwitterETL:
             objects that replicate what we get from the extract method.
         """
         mocker.patch(
-            "socialetl.socialetl.db_factory",
+            "socialetl.social_etl.db_factory",
             return_value=DatabaseConnection(db_file="data/test.db"),
         )
-        # Create a TwitterETL object
         # Create a TwitterETL object
         _, twitter_etl = etl_factory('twitter')
         # Call the transform method on the TwitterETL object
         # and pass in the mock
         transformed_data = twitter_etl.transform(
-            mock_twitter_data,
-            TransformationType('no_tx').transformation_factory(),
+            mock_twitter_data, transformation_factory('no_tx')
         )
-        # Create a DatabaseConnection with db_type as :memory:
-
         db_cursor_context = DatabaseConnection(
             db_file="data/test.db"
         ).managed_cursor()
